# src/geek/class4/coin/api/coin/v1/coin.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "api/coin/v1/coin.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "Coin"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/coin/add": {
      "post": {
        "operationId": "Coin_AddCoin",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1AddCoinReply"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1AddCoinRequest"
            }
          }
        ],
        "tags": [
          "Coin"
        ]
      }
    },
    "/v1/coin/reduce": {
      "post": {
        "operationId": "Coin_ReduceCoin",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1ReduceCoinReply"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1ReduceCoinRequest"
            }
          }
        ],
        "tags": [
          "Coin"
        ]
      }
    },
    "/v1/coin/show": {
      "get": {
        "operationId": "Coin_ShowCoin",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1ShowCoinReply"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "tags": [
          "Coin"
        ]
      }
    }
  },
  "definitions": {
    "protobufAny": {
      "type": "object",
      "properties": {
        "typeUrl": {
          "type": "string"
        },
        "value": {
          "type": "string",
          "format": "byte"
        }
      }
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
    "v1AddCoinReply": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string"
        }
      }
    },
    "v1AddCoinRequest": {
      "type": "object",
      "properties": {
        "num": {
          "type": "string",
          "format": "int64"
        }
      }
    },
    "v1ReduceCoinReply": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string"
        }
      }
    },
    "v1ReduceCoinRequest": {
      "type": "object",
      "properties": {
        "num": {
          "type": "string",
          "format": "int64"
        }
      }
    },
    "v1ShowCoinReply": {
      "type": "object",
      "properties": {
        "amount": {
          "type": "string",
          "format": "int64"
        },
        "message": {
          "type": "string"
        }
      }
    }
  }
}